requirements.txt
flask_sqlalchemy
Flask-Migrate
datetime
flask_admin
python-dotenv